Move files, add missing semicolons, remove old files
