+++ lib/file-storage-migration.js
@@ -1,18 +1,19 @@
 var util = require('util');
+var Migration = require('./migration')
 
 function FileStorageMigration(opt){
   if (typeof opt != 'object') opt = {};
   this.directory = opt.directory || path.resolve('migrations');
   this.db = opt.db;
   this.constructor.super_.call(this);
-}
+};
 
 util.inherits(FileStorageMigration, require('./migration-manager'));
 
 
 FileStorageMigration.prototype.compareMigration = function(mig1, mig2) {
   mig1.name == mig2.name;
-}
+};
 
 
 FileStorageMigration.prototype.filterNewMigrations = function() {
@@ -23,7 +24,7 @@ FileStorageMigration.prototype.filterNewMigrations = function() {
   return this.newMigrations = this.migrations.filter(function(migration) {
     return completedKeys.indexOf(migration.name) == -1;
   });
-}
+};
 
 
 FileStorageMigration.prototype.loadAllMigrations = function(callback) {
@@ -34,50 +35,15 @@ FileStorageMigration.prototype.loadAllMigrations = function(callback) {
 
   var migrations = [];
   for (var i=0; i < migrationFiles.length; ++i) {
-    var migrationFile = migrationFiles[i];
+    var migration = migrationFiles[i];
     migration = new Migration({
-      name: migrationFile,
-      path: path.join(self.directory, migrationFile)
+      name: migration,
+      path: path.join(self.directory, migration)
     });
     migrations.push(migration);
   };
   callback(null, migrations);
-}
-
-
-FileStorageMigration.prototype.loadCompletedMigrations = function(callback) {
-  fs.readFile(this.directory+'/.migrate', 'utf8', function(err, json){
-    if (err && err.code == 'ENOENT') return callback(null, []);
-    if (err) return callback(err);
-    try {
-      callback(null, JSON.parse(json));
-    } catch (err) {
-      callback(err);
-    }
-  });
-}
-
-
-var populateMigration = function(migration, callback) {
-  try {
-    migrationFile = require(migration.path);
-  } catch (err) {
-    return callback(err);
-  }
-  migration.up = migrationFile.up;
-  migration.down = migrationFile.down;
-  callback(null, migration);
-}
-
-
-FileStorageMigration.prototype.executeMigration = function(direction, migration, callback) {
-  var self = this;
-  populateMigration(migration, function(err, migration) {
-    if (err) return callback(err);
-    self.current = migration;
-    migration[direction] && migration[direction].call(self, callback);
-  })
-}
+};
 
 
 var readJson = function(file, callback) {
@@ -94,14 +60,52 @@ var readJson = function(file, callback) {
 };
 
 
+FileStorageMigration.prototype.loadCompletedMigrations = function(callback) {
+  readJson(this.directory+'/.migrate', function(err, arr){
+    if (err) return callback(err);
+    var migrations = [];
+    for (var i=0; i < arr.length; ++i) {
+      var migration = arr[i];
+      migration = new Migration({
+        name: migration.name,
+        path: path.join(self.directory, migration.name)
+      });
+      migrations.push(migration);
+    };
+  });
+};
+
+
+var populateMigration = function(migration, callback) {
+  try {
+    migrationFile = require(migration.path);
+  } catch (err) {
+    return callback(err);
+  }
+  migration.up = migrationFile.up;
+  migration.down = migrationFile.down;
+  callback(null, migration);
+};
+
+
+FileStorageMigration.prototype.executeMigration = function(direction, migration, callback) {
+  var self = this;
+  populateMigration(migration, function(err, migration) {
+    if (err) return callback(err);
+    self.current = migration;
+    migration[direction] && migration[direction].call(self, callback);
+  })
+};
+
+
 FileStorageMigration.prototype.save = function(migration, callback) {
   var self = this;
   readJson(this.directory+'/.migrate', function(err, json){
     if (err) return callback(err);
     json.push({name: migration.name});
-    fs.writeFile(this.directory+'/.migrate', json, callback)
+    fs.writeFile(this.directory+'/.migrate', JSON.stringify(json), callback)
   });
-}
+};
 
 
 FileStorageMigration.prototype.delete = function(migration, callback) {
@@ -113,6 +117,6 @@ FileStorageMigration.prototype.delete = function(migration, callback) {
     });
     fs.writeFile(this.directory+'/.migrate', JSON.stringify(json), callback)
   });
-}
+};
 
 module.exports = FileStorageMigration;
