+++ lib/migration-manager.js
@@ -9,7 +9,7 @@ function MigrationManager() {
   this.completedMigrations = [];
   this.newMigrations = [];
   this.jobList = [];
-}
+};
 
 util.inherits(MigrationManager, require('events').EventEmitter);
 
@@ -30,21 +30,16 @@ MigrationManager.prototype.migrate = function(direction, count, callback) {
     if (err) return callback(err);
     return self._migrate(direction, count, callback);
   });
-}
+};
 
 
-MigrationManager.prototype.up = function(count, callback) {
-  var args = Array.prototype.slice.call(arguments);
-  args.unshift('up');
-  this.migrate.apply(this, args);
-}
-
-
-MigrationManager.prototype.down = function(count, callback) {
-  var args = Array.prototype.slice.call(arguments);
-  args.unshift('down');
-  this.migrate.apply(this, args);
-}
+['up', 'down'].forEach(function(method) {
+  MigrationManager.prototype[method] = function(count, callback) {
+    var args = Array.prototype.slice.call(arguments);
+    args.unshift(method);
+    this.migrate.apply(this, args);
+  }
+});
 
 
 //
@@ -52,46 +47,62 @@ MigrationManager.prototype.down = function(count, callback) {
 //
 var notImplementedError = function(name) {
   throw new Error('MigrationManager.'+name+' is not implemented.');
-}
+};
+
 
 MigrationManager.prototype.compareMigration = function(mig1, mig2) {
   notImplementedError('compareMigration');
-}
+};
 
 
 MigrationManager.prototype.filterNewMigrations = function() {
   notImplementedError('filterNewMigrations');
-}
+};
 
 
 MigrationManager.prototype.loadAllMigrations = function(callback) {
   notImplementedError('loadAllMigrations');
-}
+};
 
 
 MigrationManager.prototype.loadCompletedMigrations = function(callback) {
   notImplementedError('loadCompletedMigrations');
-}
+};
 
 
 MigrationManager.prototype.save = function(migration, callback) {
   notImplementedError('save');
-}
+};
 
 
 MigrationManager.prototype.delete = function(migration, callback) {
   notImplementedError('delete');
-}
+};
 
 
 MigrationManager.prototype.executeMigration = function(direction, migration, callback) {
   notImplementedError('executeMigration');
-}
+};
 
 
 //
 // API for Modules
 //
+MigrationManager.prototype.addMigration = function(migration) {
+  this.migrations.push(migration);
+};
+
+
+MigrationManager.prototype.addCompletedMigration = function(migration) {
+  this.completedMigrations.push(migration);
+};
+
+
+MigrationManager.prototype.addJob = function(migration) {
+  this.jobList.push(migration);
+};
+
+
 MigrationManager.prototype.execute = function(direction, callback) {
   var self = this;
   if (!this.jobList.length) return callback();
@@ -105,22 +116,7 @@ MigrationManager.prototype.execute = function(direction, callback) {
         .execute(direction, callback);
     })
   });
-}
-
-
-MigrationManager.prototype.addMigration = function(migration) {
-  this.migrations.push(migration);
-}
-
-
-MigrationManager.prototype.addCompletedMigration = function(migration) {
-  this.completedMigrations.push(migration);
-}
-
-
-MigrationManager.prototype.addJob = function(migration) {
-  this.jobList.push(migration);
-}
+};
 
 
 //
@@ -136,7 +132,7 @@ MigrationManager.prototype._load = function(callback) {
       callback();
     });
   });
-}
+};
 
 
 MigrationManager.prototype._loadAllMigrations = function(callback) {
@@ -148,7 +144,7 @@ MigrationManager.prototype._loadAllMigrations = function(callback) {
     };
     callback(null);
   });
-}
+};
 
 
 MigrationManager.prototype._loadCompletedMigrations = function(callback) {
@@ -160,14 +156,14 @@ MigrationManager.prototype._loadCompletedMigrations = function(callback) {
     };
     callback(null);
   });
-}
+};
 
 
 MigrationManager.prototype._migrate = function(direction, count, callback) {
   var list = direction == 'up'? this.newMigrations : this.completedMigrations.reverse();
   this.jobList = list.slice(0, count || Infinity);
   this.execute(direction, callback);
-}
+};
 
 
 MigrationManager.prototype._shiftStack = function(direction) {
@@ -177,7 +173,7 @@ MigrationManager.prototype._shiftStack = function(direction) {
     this.newMigrations.unshift(this.completedMigrations.pop());
   }
   return this;
-}
+};
 
 
 module.exports = MigrationManager;
